Core/Warden: Update copyright information / More cleanups.
thanks to: leak

+++ src/server/game/Warden/WardenCheckMgr.cpp
@@ -1,10 +1,11 @@
 /*
- * Copyright (C) 2008-2011 TrinityCore <http://www.trinitycore.org/>
- * Copyright (C) 2005-2011 MaNGOS <http://getmangos.com/>
+ * Copyright (C) 2011-2012 Project SkyFire <http://www.projectskyfire.org/>
+ * Copyright (C) 2008-2012 TrinityCore <http://www.trinitycore.org/>
+ * Copyright (C) 2005-2012 MaNGOS <http://getmangos.com/>
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
- * Free Software Foundation; either version 2 of the License, or (at your
+ * Free Software Foundation; either version 3 of the License, or (at your
  * option) any later version.
  *
  * This program is distributed in the hope that it will be useful, but WITHOUT
@@ -177,7 +178,7 @@ void WardenCheckMgr::LoadWardenChecks()
     sLog->outString();
 }
 
-WardenCheck* WardenCheckMgr::GetWardenDataById(uint32 Id)
+WardenCheck* WardenCheckMgr::GetWardenDataById(uint16 Id)
 {
     if (Id < CheckStore.size())
         return CheckStore[Id];
@@ -185,7 +186,7 @@ WardenCheck* WardenCheckMgr::GetWardenDataById(uint32 Id)
     return NULL;
 }
 
-WardenCheckResult* WardenCheckMgr::GetWardenResultById(uint32 Id)
+WardenCheckResult* WardenCheckMgr::GetWardenResultById(uint16 Id)
 {
     CheckResultContainer::const_iterator itr = CheckResultStore.find(Id);
     if (itr != CheckResultStore.end())
