+++ src/server/game/Warden/WardenCheckMgr.h
@@ -1,10 +1,11 @@
 /*
- * Copyright (C) 2008-2011 TrinityCore <http://www.trinitycore.org/>
- * Copyright (C) 2005-2011 MaNGOS <http://getmangos.com/>
+ * Copyright (C) 2011-2012 Project SkyFire <http://www.projectskyfire.org/>
+ * Copyright (C) 2008-2012 TrinityCore <http://www.trinitycore.org/>
+ * Copyright (C) 2005-2012 MaNGOS <http://getmangos.com/>
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
- * Free Software Foundation; either version 2 of the License, or (at your
+ * Free Software Foundation; either version 3 of the License, or (at your
  * option) any later version.
  *
  * This program is distributed in the hope that it will be useful, but WITHOUT
@@ -55,8 +56,8 @@ class WardenCheckMgr
         typedef std::vector<WardenCheck*> CheckContainer;
         typedef std::map<uint32, WardenCheckResult*> CheckResultContainer;
 
-        WardenCheck* GetWardenDataById(uint32 Id);
-        WardenCheckResult* GetWardenResultById(uint32 Id);
+        WardenCheck* GetWardenDataById(uint16 Id);
+        WardenCheckResult* GetWardenResultById(uint16 Id);
 
         uint32 InternalDataID;
         std::vector<uint16> MemChecksIdPool;
